Move caches to valorless.cache and rename methods
Move PlayerCache and SkinCache classes from valorless.player to valorless.cache (file renames). Simplify PlayerCache API by renaming getUUIDByUsername -> getUUID and getPlayerByUsername -> getPlayer. Update imports in ValorlessUtils accordingly. No behavior changes intended, but external callers must be updated for the new package paths and method names.
