+++ src/main/java/valorless/cache/SkinCache.java
@@ -1,4 +1,4 @@
-package valorless.player;
+package valorless.cache;
 
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
